+++ rust-toolchain
@@ -1,1 +1,1 @@
-nightly-2018-12-06
+nightly-2019-04-19
