+++ sector-base/src/api/disk_backed_storage.rs
@@ -1,4 +1,3 @@
-use libc;
 use std::fs::{create_dir_all, remove_file, File, OpenOptions};
 use std::io::{Read, Seek, SeekFrom};
 use std::path::Path;
